backend/upload-files/requirements.txt
pydantic==2.5.0
boto3==1.34.0
psycopg2-binary==2.9.9
